Refactoring and removing reduction
