+++ lltk/__init__.py
@@ -1,1 +1,7 @@
 __version__ = '0.1-dev'
+
+from .sentencesplitter import SentenceSplitter
+from .smmry import Smmry
+from .stemmer import Stemmer
+
+__all__ = ["SentenceSplitter", "Smmry", "Stemmer"]
